refactor(test): migrate zero-run-service tests to route integration test (#9416)
