+++ turbo/apps/web/app/api/zero/runs/__tests__/route.test.ts
@@ -1,12 +1,18 @@
-import { describe, it, expect, beforeEach, vi } from "vitest";
+import { describe, it, expect, beforeEach, afterEach, vi } from "vitest";
 import { POST } from "../route";
 import {
   createTestRequest,
   createTestCompose,
   createTestSessionWithConversation,
+  createTestConnector,
+  createTestUserConnector,
   findTestZeroRun,
+  findTestRunRecord,
+  findTestRunnerJobEntry,
   insertOrgDefaultModelProvider,
+  insertUserCacheEntry,
 } from "../../../../../src/__tests__/api-test-helpers";
+import { createTestZeroAgent } from "../../../../../src/__tests__/db-test-seeders/agents";
 import { getTestZeroAgentId } from "../../../../../src/__tests__/db-test-assertions/agents";
 import {
   testContext,
@@ -17,13 +23,29 @@ import { mockClerk } from "../../../../../src/__tests__/clerk-mock";
 import {
   generateSandboxToken,
   generateZeroToken,
+  verifyZeroToken,
 } from "../../../../../src/lib/auth/sandbox-token";
+import { decryptSecretsMap } from "../../../../../src/lib/shared/crypto/secrets-encryption";
 import { reloadEnv } from "../../../../../src/env";
 import { seedTestRun } from "../../../../../src/__tests__/db-test-seeders/runs";
+import { updateUserPreferences } from "../../../../../src/lib/zero/user/user-preferences-service";
+import { updateUserFeatureSwitches } from "../../../../../src/lib/zero/user/feature-switches-service";
+import { FeatureSwitchKey } from "@vm0/core";
+import * as core from "@vm0/core";
 
 const context = testContext();
 
 const URL = "http://localhost:3000/api/zero/runs";
+
+function postRun(body: Record<string, unknown>) {
+  return POST(
+    createTestRequest(URL, {
+      method: "POST",
+      headers: { "Content-Type": "application/json" },
+      body: JSON.stringify(body),
+    }),
+  );
+}
 
 describe("POST /api/zero/runs", () => {
   beforeEach(() => {
@@ -237,6 +259,440 @@ describe("POST /api/zero/runs", () => {
       const zeroRun = await findTestZeroRun(data.runId);
       expect(zeroRun).toBeDefined();
       expect(zeroRun!.triggerAgentId).toBeNull();
+    });
+  });
+
+  describe("zero-layer defaults", () => {
+    let user: UserContext;
+    let agentId: string;
+
+    beforeEach(async () => {
+      user = await context.setupUser();
+      const agentName = uniqueId("agent");
+      await createTestCompose(agentName);
+      agentId = await getTestZeroAgentId(user.orgId, agentName);
+      vi.stubEnv("RUNNER_DEFAULT_GROUP", "vm0/production");
+      reloadEnv();
+    });
+
+    it("should inject memoryName into execution context", async () => {
+      const response = await postRun({ agentId, prompt: "Hello" });
+      expect(response.status).toBe(201);
+      const data = await response.json();
+
+      const job = await findTestRunnerJobEntry(data.runId);
+      expect(job).toBeDefined();
+      expect(job!.executionContext.memoryName).toBe("memory");
+    });
+
+    it("should not inject artifact into storage manifest", async () => {
+      const response = await postRun({ agentId, prompt: "Hello" });
+      expect(response.status).toBe(201);
+      const data = await response.json();
+
+      const job = await findTestRunnerJobEntry(data.runId);
+      expect(job).toBeDefined();
+      expect(job!.executionContext.storageManifest?.artifact).toBeNull();
+    });
+
+    it("should inject memory into storage manifest", async () => {
+      const response = await postRun({ agentId, prompt: "Hello" });
+      expect(response.status).toBe(201);
+      const data = await response.json();
+
+      const job = await findTestRunnerJobEntry(data.runId);
+      expect(job).toBeDefined();
+      expect(job!.executionContext.storageManifest).not.toBeNull();
+      expect(job!.executionContext.storageManifest!.memory).not.toBeNull();
+    });
+
+    it("should inject agent identity into appendSystemPrompt", async () => {
+      const agentName = uniqueId("identity-agent");
+      await createTestCompose(agentName);
+      await createTestZeroAgent(user.orgId, agentName, {
+        displayName: "My Agent",
+        description: "A helpful assistant",
+        sound: "friendly",
+      });
+      const identityAgentId = await getTestZeroAgentId(user.orgId, agentName);
+
+      const response = await postRun({
+        agentId: identityAgentId,
+        prompt: "Hello",
+      });
+      expect(response.status).toBe(201);
+      const data = await response.json();
+
+      const run = await findTestRunRecord(data.runId);
+      expect(run).toBeDefined();
+      expect(run!.appendSystemPrompt).toContain("My Agent");
+      expect(run!.appendSystemPrompt).toContain("A helpful assistant");
+    });
+
+    it("should inject agent tools prompt even when no identity metadata exists", async () => {
+      const response = await postRun({ agentId, prompt: "Hello" });
+      expect(response.status).toBe(201);
+      const data = await response.json();
+
+      const run = await findTestRunRecord(data.runId);
+      expect(run).toBeDefined();
+      expect(run!.appendSystemPrompt).toContain("# Agent Tools");
+      expect(run!.appendSystemPrompt).not.toContain("# Agent Identity");
+    });
+
+    it("should inject ZERO_TOKEN into execution context secrets", async () => {
+      const response = await postRun({ agentId, prompt: "Hello" });
+      expect(response.status).toBe(201);
+      const data = await response.json();
+
+      const job = await findTestRunnerJobEntry(data.runId);
+      expect(job).toBeDefined();
+      const encrypted = job!.executionContext.encryptedSecrets;
+      expect(encrypted).not.toBeNull();
+
+      const secrets = decryptSecretsMap(
+        encrypted,
+        globalThis.services.env.SECRETS_ENCRYPTION_KEY,
+      );
+      expect(secrets).not.toBeNull();
+      expect(secrets!.ZERO_TOKEN).toBeDefined();
+
+      const auth = verifyZeroToken(secrets!.ZERO_TOKEN!);
+      expect(auth).not.toBeNull();
+      expect(auth!.userId).toBe(user.userId);
+      expect(auth!.runId).toBe(data.runId);
+      expect(auth!.orgId).toBe(user.orgId);
+      expect(auth!.capabilities).toEqual(
+        expect.arrayContaining(["agent:read", "agent:write"]),
+      );
+    });
+
+    it("should inject disallowedTools with cron tools", async () => {
+      const response = await postRun({ agentId, prompt: "Hello" });
+      expect(response.status).toBe(201);
+      const data = await response.json();
+
+      const job = await findTestRunnerJobEntry(data.runId);
+      expect(job).toBeDefined();
+      expect(job!.executionContext.disallowedTools).toEqual(
+        expect.arrayContaining(["CronCreate", "CronList", "CronDelete"]),
+      );
+    });
+  });
+
+  describe("parameter forwarding", () => {
+    let user: UserContext;
+    let agentId: string;
+
+    beforeEach(async () => {
+      user = await context.setupUser();
+      const agentName = uniqueId("fwd-agent");
+      await createTestCompose(agentName);
+      agentId = await getTestZeroAgentId(user.orgId, agentName);
+      vi.stubEnv("RUNNER_DEFAULT_GROUP", "vm0/production");
+      reloadEnv();
+    });
+
+    it("should leave continuedFromSessionId null when no sessionId given", async () => {
+      const response = await postRun({ agentId, prompt: "Hello" });
+      expect(response.status).toBe(201);
+      const data = await response.json();
+
+      const run = await findTestRunRecord(data.runId);
+      expect(run).toBeDefined();
+      expect(run!.continuedFromSessionId).toBeNull();
+    });
+  });
+
+  describe("user info injection", () => {
+    let user: UserContext;
+    let agentId: string;
+
+    beforeEach(async () => {
+      user = await context.setupUser();
+      const agentName = uniqueId("info-agent");
+      await createTestCompose(agentName);
+      agentId = await getTestZeroAgentId(user.orgId, agentName);
+      vi.stubEnv("RUNNER_DEFAULT_GROUP", "vm0/production");
+      reloadEnv();
+    });
+
+    it("should inject # Current User Info with email and timezone into appendSystemPrompt", async () => {
+      const response = await postRun({ agentId, prompt: "Hello" });
+      expect(response.status).toBe(201);
+      const data = await response.json();
+
+      const run = await findTestRunRecord(data.runId);
+      expect(run).toBeDefined();
+      expect(run!.appendSystemPrompt).toContain("# Current User Info");
+      expect(run!.appendSystemPrompt).toContain("Email:");
+      expect(run!.appendSystemPrompt).toContain("Timezone:");
+    });
+
+    it("should default timezone to UTC when user has no timezone preference", async () => {
+      const response = await postRun({ agentId, prompt: "Hello" });
+      expect(response.status).toBe(201);
+      const data = await response.json();
+
+      const run = await findTestRunRecord(data.runId);
+      expect(run).toBeDefined();
+      expect(run!.appendSystemPrompt).toContain("Timezone: UTC");
+    });
+
+    it("should use user timezone preference when set", async () => {
+      await updateUserPreferences(user.orgId, user.userId, {
+        timezone: "Asia/Shanghai",
+      });
+
+      const response = await postRun({ agentId, prompt: "Hello" });
+      expect(response.status).toBe(201);
+      const data = await response.json();
+
+      const run = await findTestRunRecord(data.runId);
+      expect(run).toBeDefined();
+      expect(run!.appendSystemPrompt).toContain("Timezone: Asia/Shanghai");
+    });
+
+    it("should include cached user name when available", async () => {
+      await insertUserCacheEntry({
+        userId: user.userId,
+        email: "alice@example.com",
+        name: "Alice Zhang",
+      });
+
+      const response = await postRun({ agentId, prompt: "Hello" });
+      expect(response.status).toBe(201);
+      const data = await response.json();
+
+      const run = await findTestRunRecord(data.runId);
+      expect(run).toBeDefined();
+      expect(run!.appendSystemPrompt).toContain("Name: Alice Zhang");
+    });
+
+    it("should omit name from user info when not cached", async () => {
+      const response = await postRun({ agentId, prompt: "Hello" });
+      expect(response.status).toBe(201);
+      const data = await response.json();
+
+      const run = await findTestRunRecord(data.runId);
+      expect(run).toBeDefined();
+      expect(run!.appendSystemPrompt).toContain("# Current User Info");
+      expect(run!.appendSystemPrompt).not.toContain("Name:");
+    });
+  });
+
+  describe("permission policies", () => {
+    let user: UserContext;
+
+    beforeEach(async () => {
+      user = await context.setupUser();
+      vi.stubEnv("RUNNER_DEFAULT_GROUP", "vm0/production");
+      reloadEnv();
+    });
+
+    it("should carry all permissions and grant only allowed ones", async () => {
+      const agentName = uniqueId("fw-agent");
+      await createTestCompose(agentName);
+      await createTestConnector({ type: "slack" });
+      await createTestZeroAgent(user.orgId, agentName, {
+        permissionPolicies: {
+          slack: {
+            "channels:read": "allow",
+            "channels:history": "allow",
+            admin: "deny",
+          },
+        },
+      });
+      const agentId = await getTestZeroAgentId(user.orgId, agentName);
+      await createTestUserConnector(user.orgId, user.userId, agentId, "slack");
+
+      const response = await postRun({ agentId, prompt: "Hello" });
+      expect(response.status).toBe(201);
+      const data = await response.json();
+
+      const job = await findTestRunnerJobEntry(data.runId);
+      expect(job).toBeDefined();
+      const firewalls = job!.executionContext.firewalls;
+      expect(firewalls).toBeDefined();
+      const slackFirewall = firewalls!.find((fw) => {
+        return fw.ref === "slack";
+      });
+      expect(slackFirewall).toBeDefined();
+      const permNames = slackFirewall!.apis[0]!.permissions!.map((p) => {
+        return p.name;
+      });
+      expect(permNames).toContain("channels:read");
+      expect(permNames).toContain("channels:history");
+      expect(permNames).toContain("admin");
+
+      const granted = job!.executionContext.networkPolicies;
+      expect(granted).toBeDefined();
+      const slackGrant = granted!.slack;
+      expect(slackGrant).toBeDefined();
+      const grantedPerms = slackGrant!.allow;
+      expect(grantedPerms).toContain("channels:read");
+      expect(grantedPerms).toContain("channels:history");
+      expect(grantedPerms).not.toContain("admin");
+      expect(slackGrant!.unknownPolicy).toBe("allow");
+    });
+
+    it("should grant all permissions when no explicit policies exist", async () => {
+      const agentName = uniqueId("fw-nopol");
+      await createTestCompose(agentName);
+      await createTestConnector({ type: "slack" });
+      await createTestZeroAgent(user.orgId, agentName, {});
+      const agentId = await getTestZeroAgentId(user.orgId, agentName);
+      await createTestUserConnector(user.orgId, user.userId, agentId, "slack");
+
+      const response = await postRun({ agentId, prompt: "Hello" });
+      expect(response.status).toBe(201);
+      const data = await response.json();
+
+      const job = await findTestRunnerJobEntry(data.runId);
+      expect(job).toBeDefined();
+      const firewalls = job!.executionContext.firewalls;
+      expect(firewalls).toBeDefined();
+      const slackFw = firewalls!.find((fw) => {
+        return fw.ref === "slack";
+      });
+      expect(slackFw).toBeDefined();
+
+      const granted = job!.executionContext.networkPolicies;
+      expect(granted).toBeDefined();
+      const grantedPerms = granted!.slack!.allow;
+      expect(Array.isArray(grantedPerms)).toBe(true);
+      expect(grantedPerms).toContain("channels:read");
+      expect(grantedPerms).toContain("users:read");
+      expect(grantedPerms).not.toContain("admin");
+      expect(grantedPerms).not.toContain("chat:write");
+    });
+
+    it("should merge partial stored denies with default policies", async () => {
+      const agentName = uniqueId("fw-allden");
+      await createTestCompose(agentName);
+      await createTestConnector({ type: "slack" });
+      await createTestZeroAgent(user.orgId, agentName, {
+        permissionPolicies: {
+          slack: { "bookmarks:read": "deny", "bookmarks:write": "deny" },
+        },
+      });
+      const agentId = await getTestZeroAgentId(user.orgId, agentName);
+      await createTestUserConnector(user.orgId, user.userId, agentId, "slack");
+
+      const response = await postRun({ agentId, prompt: "Hello" });
+      expect(response.status).toBe(201);
+      const data = await response.json();
+
+      const job = await findTestRunnerJobEntry(data.runId);
+      expect(job).toBeDefined();
+      const slackFw = job!.executionContext.firewalls?.find((fw) => {
+        return fw.ref === "slack";
+      });
+      expect(slackFw).toBeDefined();
+      expect(slackFw!.apis[0]!.permissions!.length).toBeGreaterThan(0);
+
+      const granted = job!.executionContext.networkPolicies;
+      expect(granted).toBeDefined();
+      expect(granted!.slack!.allow).not.toContain("bookmarks:read");
+      expect(granted!.slack!.allow).not.toContain("bookmarks:write");
+      expect(granted!.slack!.allow).toContain("channels:read");
+      expect(granted!.slack!.allow).toContain("users:read");
+      expect(granted!.slack!.unknownPolicy).toBe("allow");
+    });
+
+    it("should add multiple firewall entries for multi-ref connector", async () => {
+      const agentName = uniqueId("fw-multi");
+      await createTestCompose(agentName);
+      await createTestConnector({ type: "github" });
+      await createTestConnector({ type: "slack" });
+      const agentId = await getTestZeroAgentId(user.orgId, agentName);
+      await createTestUserConnector(user.orgId, user.userId, agentId, "github");
+      await createTestUserConnector(user.orgId, user.userId, agentId, "slack");
+
+      const response = await postRun({ agentId, prompt: "Hello" });
+      expect(response.status).toBe(201);
+      const data = await response.json();
+
+      const job = await findTestRunnerJobEntry(data.runId);
+      expect(job).toBeDefined();
+      const firewalls = job!.executionContext.firewalls;
+      expect(firewalls).toBeDefined();
+      const ghFw = firewalls!.find((fw) => {
+        return fw.ref === "github";
+      });
+      const slackFw = firewalls!.find((fw) => {
+        return fw.ref === "slack";
+      });
+      expect(ghFw).toBeDefined();
+      expect(slackFw).toBeDefined();
+    });
+  });
+
+  describe("AutoSkill guidance injection", () => {
+    let user: UserContext;
+    let agentId: string;
+
+    beforeEach(async () => {
+      user = await context.setupUser();
+      const agentName = uniqueId("skill-agent");
+      await createTestCompose(agentName);
+      agentId = await getTestZeroAgentId(user.orgId, agentName);
+      vi.stubEnv("RUNNER_DEFAULT_GROUP", "vm0/production");
+      reloadEnv();
+    });
+
+    afterEach(() => {
+      vi.restoreAllMocks();
+    });
+
+    it("should inject skill guidance when AutoSkill feature switch is enabled", async () => {
+      vi.spyOn(core, "isFeatureEnabled").mockImplementation(
+        (key: FeatureSwitchKey) => {
+          if (key === FeatureSwitchKey.AutoSkill) return true;
+          return false;
+        },
+      );
+
+      const response = await postRun({ agentId, prompt: "Hello" });
+      expect(response.status).toBe(201);
+      const data = await response.json();
+
+      const run = await findTestRunRecord(data.runId);
+      expect(run).toBeDefined();
+      expect(run!.appendSystemPrompt).toContain("# Skill Management Guidance");
+      expect(run!.appendSystemPrompt).toContain("zero skill create");
+    });
+
+    it("should not inject skill guidance when AutoSkill feature switch is disabled", async () => {
+      const response = await postRun({ agentId, prompt: "Hello" });
+      expect(response.status).toBe(201);
+      const data = await response.json();
+
+      const run = await findTestRunRecord(data.runId);
+      expect(run).toBeDefined();
+      expect(run!.appendSystemPrompt).not.toContain(
+        "# Skill Management Guidance",
+      );
+    });
+
+    it("should pass user overrides to AutoSkill feature check", async () => {
+      const spy = vi.spyOn(core, "isFeatureEnabled");
+
+      await updateUserFeatureSwitches(user.orgId, user.userId, {
+        [FeatureSwitchKey.AutoSkill]: false,
+      });
+
+      const response = await postRun({ agentId, prompt: "Hello" });
+      expect(response.status).toBe(201);
+
+      expect(spy).toHaveBeenCalledWith(
+        FeatureSwitchKey.AutoSkill,
+        expect.objectContaining({
+          overrides: expect.objectContaining({
+            [FeatureSwitchKey.AutoSkill]: false,
+          }),
+        }),
+      );
     });
   });
 });
